feat: restructure tasks
