+++ api/src/utils.ts
@@ -1,0 +1,4 @@
+/**
+ * First day of stacks 2.0
+ */
+export const startDate = new Date(2021, 0, 14);
